requirements.txt
# -e git://github.com/ZombieToof/django-phpbb.git
configobj==5.0.6
django==1.11.4
django-braces==1.11.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-filter==1.0.4
django-model-utils==3.0.0
django-precise-bbcode==1.2.8
djangorestframework==3.6.3
path.py==10.3.1
postmarkup==1.2.2
pylint
